Switched default libsql driver initializer to new initializer type, adapted tests
